create_author -> that takes the author schema and adds it tothe db
